// app/components/my-account/ChannelsTable/types.ts
export type Channel = {
  id: string | number;
  channelId: string;
  title: string;
  priceClick1: number;
  priceClick2: number;
  priceClick3: number;
  priceClick4: number;
  priceClick5: number;
  priceClick6: number;
  pricePDP1: number;
  pricePDP2: number;
  pricePDP3: number;
  pricePDP4: number;
  pricePDP5: number;
  pricePDP6: number;
};